+++ client/client_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/komuw/ong/log"
 
-	"github.com/akshayjshah/attest"
+	"go.akshayshah.org/attest"
 	"go.uber.org/goleak"
 	"golang.org/x/exp/slog"
 )