+++ middleware/loadshed_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/akshayjshah/attest"
+	"go.akshayshah.org/attest"
 )
 
 const loadShedderTestHeader = "LoadShedderTestHeader"