+++ middleware/log_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/komuw/ong/id"
 	"github.com/komuw/ong/log"
 
-	"github.com/akshayjshah/attest"
+	"go.akshayshah.org/attest"
 	"golang.org/x/exp/slog"
 )