+++ middleware/recoverer_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/komuw/ong/errors"
 	"github.com/komuw/ong/log"
 
-	"github.com/akshayjshah/attest"
+	"go.akshayshah.org/attest"
 	"golang.org/x/exp/slog"
 )