+++ server/pprof_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/komuw/ong/log"
 
-	"github.com/akshayjshah/attest"
+	"go.akshayshah.org/attest"
 )
 
 func TestPprofServer(t *testing.T) {